Working on Music Player

+++ app/src/main/java/lauzhack2018/music/music_runner/PlayerAdapter.java
@@ -1,7 +1,7 @@
 package lauzhack2018.music.music_runner;
 
 public interface PlayerAdapter {
-    void loadSong(int id);
+    void loadSong(long id);
     void release();
     boolean isPlaying();
     void play();
